Move test setup to helper
