+++ tests/LivingDocumentation.Analyzer.Tests/ForEachTests.cs
@@ -1,9 +1,5 @@
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Collections.Generic;
-using System.Linq;
 using FluentAssertions;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace LivingDocumentation.Analyzer.Tests
 {
@@ -28,7 +24,7 @@ namespace LivingDocumentation.Analyzer.Tests
             ";
 
             // Act
-            var types = VisitSyntaxTree(source);
+            var types = TestHelper.VisitSyntaxTree(source);
 
             // Assert
             types[0].Methods[0].Statements[0].Should().BeOfType<ForEach>();
@@ -53,34 +49,11 @@ namespace LivingDocumentation.Analyzer.Tests
             ";
 
             // Act
-            var types = VisitSyntaxTree(source);
+            var types = TestHelper.VisitSyntaxTree(source);
 
             // Assert
             var forEach = types[0].Methods[0].Statements[0];
             forEach.Statements.Should().HaveCount(1);
-        }
-
-        private static IReadOnlyList<TypeDescription> VisitSyntaxTree(string source)
-        {
-            source.Should().NotBeNullOrWhiteSpace("without source code there is nothing to test");
-
-            var syntaxTree = CSharpSyntaxTree.ParseText(source.Trim());
-            var compilation = CSharpCompilation.Create("Test")
-                                               .WithOptions(new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary))
-                                               .AddReferences(MetadataReference.CreateFromFile(typeof(object).Assembly.Location))
-                                               .AddSyntaxTrees(syntaxTree);
-
-            var diagnostics = compilation.GetDiagnostics();
-            diagnostics.Should().HaveCount(0, "there shoudn't be any compile errors");
-
-            var semanticModel = compilation.GetSemanticModel(syntaxTree, true);
-
-            var types = new List<TypeDescription>();
-
-            var visitor = new SourceAnalyzer(semanticModel, types, Enumerable.Empty<AssemblyIdentity>().ToList());
-            visitor.Visit(syntaxTree.GetRoot());
-
-            return types;
         }
     }
 }
